fix MSVC errors and warnings
- strcasecmp is __stricmp in windows
- toml::array::for_each with if constexpr causes MSVC to fail evaluating noexcept specifiers for visit(), use for auto instead.
- engine functions defined in cpp are declared outside extern C blocks
- size_t to int narrowing is explicit where intentional.
- use %zu size_t printf specifier

+++ src/legacy_apps/sortsub.c
@@ -46,7 +46,7 @@ BOOL dokey;
   if (nline <= 0) {
     free(buf0); free(key); return 1;
   }
-  ntot = strlen(buf0); ndline = ntot + 1;
+  ntot = (int)strlen(buf0); ndline = ntot + 1;
   if (ndline < 81)
     ndline = 81;
   if (!dokey) {
@@ -132,7 +132,7 @@ int nlim;
   for (;;) {
     if (fgets(tbuf, ndline, finp) == NULL)
       break;
-    i = strlen(tbuf) - 1;
+    i = (int)strlen(tbuf) - 1;
     if (tbuf[i] == '\n') {
       tbuf[i] = '\0';
     } else {
